Convert cypress tests to typescript
