--- cypress/fixtures/comments/index.js
@@ -1,5 +1,0 @@
-const testComment = require('./testComment');
-
-module.exports = [
-  testComment,
-];
