--- cypress/fixtures/comments/testComment.js
@@ -1,44 +1,0 @@
-module.exports = {
-    _id: "test-seeded-comment",
-    postId: "test-seeded-post",
-    createdAt: new Date(),
-    postedAt: new Date(),
-    lastSubthreadActivity: new Date(),
-    contents: {
-        originalContents: {
-            type: "ckEditorMarkup",
-            data: "Test seeded comment",
-        },
-        updateType: "initial",
-        commitMessage: "",
-        html: "Test seeded comment",
-        version: "1.0.0",
-        userId: "test-user",
-        wordCount: 3,
-    },
-    userId: "test-user",
-    author: "Test User",
-    authorIsUnreviewed: false,
-    answer: false,
-    descendentCount: 0,
-    shortform: false,
-    postVersion: "1.0.0",
-    hideKarma: false,
-    schemaVersion: 1,
-    legacy: false,
-    retracted: false,
-    deleted: false,
-    deletedPublic: false,
-    spam: false,
-    hideAuthor: false,
-    moderatorHat: false,
-    baseScore: 1,
-    score: 0.149131837436287,
-    inactive: false,
-    af: false,
-    referrer: "http://localhost:3000/posts/test-post/test-post",
-    userAgent: "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Cypress/8.3.0 Chrome/91.0.4472.124 Electron/13.1.7 Safari/537.36",
-    contents_latest: "B5n47EvitY4zqSid3",
-    voteCount: 1,
-    needsReview: true,
-}
