--- cypress/fixtures/conversations/index.js
@@ -1,5 +1,0 @@
-const testConversation = require('./testConversation');
-
-module.exports = [
-  testConversation,
-];
