--- cypress/fixtures/conversations/testConversation.js
@@ -1,12 +1,0 @@
-module.exports = {
-    _id : "test-seeded-conversation",
-    createdAt: new Date(),
-    latestActivity: new Date(),
-    participantIds : [ 
-        "test-other-user",
-        "test-user",
-    ],
-    schemaVersion : 1,
-    messageCount : 1,
-    archivedByIds : [],
-}
