--- cypress/fixtures/messages/index.js
@@ -1,5 +1,0 @@
-const testMessage = require('./testMessage');
-
-module.exports = [
-  testMessage,
-];
