--- cypress/fixtures/posts/index.js
@@ -1,7 +1,0 @@
-const testPost = require('./testPost');
-const testAdminPost = require('./testAdminPost');
-
-module.exports = [
-  testPost,
-  testAdminPost,
-];
