--- cypress/fixtures/posts/testAdminPost.js
@@ -1,66 +1,0 @@
-module.exports = {
-    url: null,
-    title: "Test Seeded Admin Post Title",
-    submitToFrontpage: true,
-    draft: false,
-    meta: false,
-    isEvent: false,
-    createdAt: new Date(),
-    postedAt: new Date(),
-    modifiedAt: new Date(),
-    frontpageDate: new Date(),
-    types: [],
-    moderationStyle: null,
-    contents: {
-        originalContents: {
-            type: "ckEditorMarkup",
-            data: "Test seeded admin post body"
-        },
-        updateType: "initial",
-        commitMessage: "",
-        html: "Test seeded admin post body",
-        version: "1.0.0",
-        userId: "test-admin",
-        wordCount: 5,
-    },
-    userId: "test-admin",
-    slug: "test-seeded-admin-post",
-    deletedDraft: false,
-    status: 2,
-    isFuture: false,
-    sticky: false,
-    stickyPriority: 2,
-    question: false,
-    authorIsUnreviewed: false,
-    hiddenRelatedQuestion: false,
-    shortform: false,
-    noIndex: false,
-    schemaVersion: 1,
-    baseScore: 1,
-    score: 0.0187169997200909,
-    inactive: false,
-    unlisted: false,
-    disableRecommendation: false,
-    defaultRecommendation: false,
-    hideFrontpageComments: false,
-    maxBaseScore: 1,
-    scoreExceeded2Date: null,
-    scoreExceeded30Date: null,
-    scoreExceeded45Date: null,
-    scoreExceeded75Date: null,
-    scoreExceeded125Date: null,
-    scoreExceeded200Date: null,
-    onlineEvent: false,
-    metaSticky: false,
-    hideAuthor: false,
-    hideCommentKarma: false,
-    af: false,
-    afSticky: false,
-    referrer: "http://localhost:3000/newPost",
-    userAgent: "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.159 Safari/537.36",
-    pingbacks: {},
-    voteCount: 1,
-    socialPreviewImageAutoUrl: "",
-    tagRelevance: {},
-    viewCount: 4
-}
