--- cypress/fixtures/users/index.js
@@ -1,11 +1,0 @@
-const testUser = require('./testUser');
-const testOtherUser = require('./testOtherUser');
-const testAdmin = require('./testAdmin');
-const testUserUnsetUsername = require('./testUserUnsetUsername');
-
-module.exports = [
-  testUser,
-  testOtherUser,
-  testAdmin,
-  testUserUnsetUsername,
-];
