+++ e2e/cypress/fixtures/comments/index.ts
@@ -1,0 +1,5 @@
+import { testComment } from './testComment';
+
+export default [
+  testComment,
+];
