+++ e2e/cypress/fixtures/conversations/index.ts
@@ -1,0 +1,5 @@
+import { testConversation } from './testConversation';
+
+export default [
+  testConversation,
+];
