+++ e2e/cypress/fixtures/conversations/testConversation.ts
@@ -1,0 +1,12 @@
+export const testConversation = {
+  _id : "test-seeded-conversation",
+  createdAt: new Date(),
+  latestActivity: new Date(),
+  participantIds : [ 
+    "test-other-user",
+    "test-user",
+  ],
+  schemaVersion : 1,
+  messageCount : 1,
+  archivedByIds : [],
+}
