+++ e2e/cypress/fixtures/messages/index.ts
@@ -1,0 +1,5 @@
+import { testMessage } from './testMessage';
+
+export default [
+  testMessage,
+];
