+++ e2e/cypress/fixtures/posts/index.ts
@@ -1,0 +1,7 @@
+import { testPost } from './testPost';
+import { testAdminPost } from './testAdminPost';
+
+export default [
+  testPost,
+  testAdminPost,
+];
