+++ e2e/cypress/fixtures/users/index.ts
@@ -1,0 +1,11 @@
+import { testUser } from './testUser';
+import { testOtherUser } from './testOtherUser';
+import { testAdmin } from './testAdmin';
+import { testUserUnsetUsername } from './testUserUnsetUsername';
+
+export default [
+  testUser,
+  testOtherUser,
+  testAdmin,
+  testUserUnsetUsername,
+];
