+++ e2e/package.json
@@ -1,0 +1,4 @@
+{
+  "type": "module",
+  "private": true
+}
