+++ package.json
@@ -30,8 +30,8 @@
     "ea-start-staging-db": "./build.js -run --watch --mongoUrlFile ../ForumCredentials/staging-db-conn.txt --settings ../ForumCredentials/settings-local-dev-staging-db.json",
     "ea-start-prod-db": "./build.js -run --watch --mongoUrlFile ../ForumCredentials/prod-db-conn.txt --settings ../ForumCredentials/settings-local-dev-prod-db.json",
     "ea-start-testing-db": "./build.js -run --mongoUrl ${TESTING_DB_URL:=$(cat ../ForumCredentials/testing-db-conn.txt)} --settings ./settings-test.json",
-    "ea-cypress-run": "yarn cypress run --env TESTING_DB_URL=${TESTING_DB_URL:=$(cat ../ForumCredentials/testing-db-conn.txt)}",
-    "ea-cypress-open": "yarn cypress open --env TESTING_DB_URL=${TESTING_DB_URL:=$(cat ../ForumCredentials/testing-db-conn.txt)}",
+    "ea-cypress-run": "yarn cypress run --project e2e --env TESTING_DB_URL=${TESTING_DB_URL:=$(cat ../ForumCredentials/testing-db-conn.txt)}",
+    "ea-cypress-open": "yarn cypress open --project e2e --env TESTING_DB_URL=${TESTING_DB_URL:=$(cat ../ForumCredentials/testing-db-conn.txt)}",
     "ea-mongod": "mongod --dbpath .meteor/local/db --port 3001",
     "ea-mongo": "mongo $(< ../ForumCredentials/dev-db-conn.txt)",
     "ea-mongo-staging": "mongo $(< ../ForumCredentials/staging-db-conn.txt)",
@@ -272,6 +272,7 @@
     "@babel/preset-env": "7.11.5",
     "@babel/preset-react": "7.10.4",
     "@babel/preset-typescript": "^7.12.1",
+    "@cypress/webpack-preprocessor": "^5.12.2",
     "@graphql-codegen/cli": "^1.9.1",
     "@graphql-codegen/fragment-matcher": "^1.9.0",
     "@graphql-codegen/introspection": "^1.9.0",
@@ -282,7 +283,7 @@
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "cross-env": "^7.0.3",
-    "cypress": "^10.4.0",
+    "cypress": "^10.8.0",
     "esbuild": "^0.15.7",
     "esbuild-jest": "^0.5.0",
     "esbuild-loader": "^2.6.1",
@@ -294,9 +295,11 @@
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^27.3.1",
+    "local-cypress": "^1.2.6",
     "mongodb-memory-server": "7.3.6",
     "ts-jest": "^27.0.7",
-    "ts-loader": "^8.0.5"
+    "ts-loader": "^9.3.1",
+    "webpack": "^5.74.0"
   },
   "resolutions": {
     "graphql": "^14.5.8",
